+++ src/firebase/authenticationFirebase.js
@@ -23,10 +23,9 @@ const displayUserData = () => {
       if (user) {
         localStorage.setItem('UserCredentialFb', JSON.stringify(user.email));
         const contentGretting = document.querySelector('#wallOffPublication');
-        console.log(contentGretting);
         if (contentGretting !== null) {
           contentGretting.innerHTML += `<h1 id="showUserEmailId" class="greetingUser">Usuario en sesión: ${user.email}</h1>`;
-          eventSignOut();
+         // eventSignOut();
         }
       } else {
         document.querySelector('#wallOffPublication').innerHTML += `<h1 id="showUserEmailId" class="greetingUser">Hola resgistrate en nuestra red social</h1>`;
@@ -35,37 +34,22 @@ const displayUserData = () => {
   }
 };
 
-const eventLoginButton = () => {
-  if (window.location.pathname === '/userSignIn') {
-    const btnLogin = document.querySelector('#submitContinue');
-    if (btnLogin) {
-      btnLogin.addEventListener('click', () => {
-        const email = document.querySelector('#userSi').value;
-        const password = document.querySelector('#passwordSi').value;
-        signInWithEmailAndPassword(auth, email, password)
-          .then((userCredential) => {
-            console.log(userCredential, 'se cumplió');
-            const greetingUser = document.getElementById('greetingUserId');
-            if (greetingUser) {
-              greetingUser.innerHTML = '';
-            }
-            window.location.href = '#wall';
-          })
-          .catch((error) => {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-          });
-      });
-    }
-  }
-};
-
+const signIn = (auth, email, password) => {
+  return signInWithEmailAndPassword(auth, email, password)
+    .then((userCredential) => {
+      const user = userCredential;
+      console.log(user)      
+    })
+    .catch((error) => {
+      const errorCode = error.code;
+      const errorMessage = error.message;
+    });
+}
+  
 const logOut = () => {
   return signOut(auth).then(() => {
     auth.signOut();
-    console.log('Ha salido de la sesión');
-
-  }).catch((error) => {
+   }).catch((error) => {
     console.error('Ha ocurrido un error al intentar salir', error);
   });
 }
@@ -97,5 +81,5 @@ export {
   eventSignOut,
   logOut,
   displayUserData,
-  eventLoginButton,
+  signIn,
 };
